+++ cogs/new_release.py
@@ -2,14 +2,15 @@ from datetime import date
 
 import discord
 from discord.ext import commands
+from assets import urls
 
 from config.read_configs import ReadConfigs as configs
 from harbinger import Harbinger
 
 CUSTOM_COLOR = configs.custom_color()
-source_code = "https://github.com/notoriouslogank/harbinger/releases"
-issues = "https://github.com/notoriouslogank/harbinger/issues"
-footer = "For information about new commands, say !help <command>."
+source_code = urls.SOURCE_CODE
+issues = urls.ISSUES
+footer = urls.FOOTER
 
 
 class Release(commands.Cog):
@@ -17,12 +18,7 @@ class Release(commands.Cog):
         self.bot = bot
         self.version = Harbinger.get_ver()
 
-    @commands.command()
-    async def release(self, ctx, text):
-        await ctx.channel.purge(limit=1)
-        await ctx.send(embed=self.release_embed(ctx, text))
-
-    def release_embed(self, ctx, text):
+    def release_embed(self, text):
         embed = discord.Embed(
             title="Harbinger: Release Notes",
             description=f"v{self.version}",
@@ -38,6 +34,11 @@ class Release(commands.Cog):
         embed.add_field(name="Report a Bug", value=f"{issues}", inline=False)
         embed.set_footer(text=f"{footer}")
         return embed
+
+    @commands.command()
+    async def release(self, ctx: commands.Context, text):
+        await ctx.channel.purge(limit=1)
+        await ctx.send(embed=self.release_embed(ctx, text))
 
 
 async def setup(bot):