+++ docs/CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.4.0] - 2024-04-17
+
+### Added
+
+- !backmc command to create a backup of the Minecraft server
+- Various support functions for !backmc in cogs/minecraft.py
+
+### Changed
+
+- Minor formatting throughout
+- Removed unused imports
+- Moved various Github urls used in embeds to assets/urls
+
 ## [5.3.0] - 2024-04-14
 
 ### Added